bentomlrunner added into serviceclasses
